Separate engine process entities from UCI logic
